[feat/#42]: 메인페이지 및 가입신청 구현
