+++ components/common/Navigation/Navigation.module.css
@@ -110,5 +110,5 @@
   height: 20px;
   margin-left: -10px;
   margin-top: -10px;
-  background-image: url(/icons/ico_payment.svg);
+  background-image: url(/icons/ico_agits_create.svg);
 }
