+++ components/common/SelectFilter/SelectFilter.module.css
@@ -36,6 +36,10 @@
 .select.header > button::before {
   border-top: 7px solid #000;
 }
+.select.header > ul {
+  left: 20px;
+  top: 46px;
+}
 .select > ul button.active {
   background-color: var(--gray-br);
   font-weight: 700;
